Change input type to float for currency calculator

+++ 01-Basics/Currency_calculator.py
@@ -1,6 +1,6 @@
-pesos = int(input('Colombian pesos:' ))
-soles = int(input('Peruvian soles:' ))
-reais = int(input('Brazillian reais:' ))
+pesos = float(input('Colombian pesos:' ))
+soles = float(input('Peruvian soles:' ))
+reais = float(input('Brazillian reais:' ))
 
 Total = pesos * 0.00028 + soles * 0.29 + reais * 0.20
 print(Total)
